refactor: move controller to own rest package
