+++ examples/process-example/process-example-core/src/main/java/io/miragon/miranum/examples/process/ProcessExampleConfiguration.java
@@ -1,7 +1,7 @@
 package io.miragon.miranum.examples.process;
 
 import io.miragon.miranum.connect.process.application.port.out.StartProcessPort;
-import io.miragon.miranum.examples.process.adapter.in.StartProcessController;
+import io.miragon.miranum.examples.process.adapter.in.rest.StartProcessController;
 import io.miragon.miranum.examples.process.application.port.in.SendMessageUseCase;
 import io.miragon.miranum.examples.process.application.service.SendMessageService;
 import org.springframework.context.annotation.Bean;
